revert renaming of platform.h only .cpp file needs different name to avoid clash with exitting file in Mbed add AnalogIn for VTarget reading
